requirements.txt
fastapi
uvicorn[standard]
tabula-py
httpx
aiofiles
pydantic
python-multipart
ics
git+https://github.com/mytja/gimsisapi
git+https://github.com/mytja/lopolis-api
sqlalchemy
asyncpg
bcrypt
pycryptodome
emoji

